Add signed overflow operation
The main program has to be compiled and linked with '-fno-sanitize-recover'
in order for the program abort to complete.

For some reason, the fuzz tester does not require this option be
specified for the main libraries.  SPECULATION: In the case of the
fuzz tester, the main routine is the fuzz library, which seems to have
this already compiled or otherwise catches undefined behaviour traps.