Set SANITIZER_UNDEFINED_BEHAVIOR_DEFAULT to OFF
This is for exploring coverage analysis. When UB SAN is ON,
clang++ specifies that every twos-complement integer op has a branch,
distinguishing the regular and overflow cases.  This reduces the coverage
count on every integer operation, which typically never overflow.

Setting the default OFF is simply an expedient for exploring how to
get better test coverage.