Try requring g++-10 as a package
Lefticus's json2cpp project uses this package:

https://github.com/lefticus/json2cpp/blob/main/.lgtm.yml

It is unclear from the logs whether this actually forces LGTM
to build with g++10.  Some log messages indicate that gcc 9.x
is in fact used.  The CMakeLists.txt file specifies C++ 17

https://github.com/lefticus/json2cpp/blob/main/CMakeLists.txt

which might be compilable with gcc 9.x.

On the other hand, this project requires the <concepts> header,
which is present in gcc 10 but not 9.  So this project build
will test which version of gcc is used for the build.