Add dedicated codecov Action
This Action is for experimentation only.  I do not expect
it to be retained.  Though it might prove to be a better approach
to code coverage than integrating it into the CI, in which case
I might retain it.

The step in the CI Action that publishes code coverage is
commented out.