Improve coverage statistics
1. It seems that the dynamic_projection_options() macro uses
the OPT_ENABLE_* command-line variables to set options---though
only when ENABLE_DEVELOPER=ON, otherwise the values are forced
and unchangeable.

2. Make CLI tests sensitive to enabled sanitizers

Some CLI tests demonstrate that a sanitizer is enabled. This commit
ensures that those same tests continue to succeed when their sanitizer
is NOT enabled, as for example during coverage analysis.

The changes check for different (or, in the case of the leak sanitizer,
absent) output when the sanitizer is disabled.

3. Specify explicit compiler releases